// .wundergraph/generated/linkbuilder.ts
interface SourceFields {
	getCityByName: {
		name: null;
		country: null;
		config: null;
	};
	getCityById: {
		id: null;
		config: null;
	};
	users: {
		distinct_on: null;
		limit: null;
		offset: null;
		order_by: null;
		where: null;
	};
	users_aggregate: {
		distinct_on: null;
		limit: null;
		offset: null;
		order_by: null;
		where: null;
	};
	users_by_pk: {
		id: null;
	};
	capsules: {
		find: null;
		limit: null;
		offset: null;
		order: null;
		sort: null;
	};
	capsulesPast: {
		find: null;
		limit: null;
		offset: null;
		order: null;
		sort: null;
	};
	capsulesUpcoming: {
		find: null;
		limit: null;
		offset: null;
		order: null;
		sort: null;
	};
	capsule: {
		id: null;
	};
	company: {};
	cores: {
		find: null;
		limit: null;
		offset: null;
		order: null;
		sort: null;
	};
	coresPast: {
		find: null;
		limit: null;
		offset: null;
		order: null;
		sort: null;
	};
	coresUpcoming: {
		find: null;
		limit: null;
		offset: null;
		order: null;
		sort: null;
	};
	core: {
		id: null;
	};
	dragons: {
		limit: null;
		offset: null;
	};
	dragon: {
		id: null;
	};
	histories: {
		find: null;
		limit: null;
		offset: null;
		order: null;
		sort: null;
	};
	historiesResult: {
		find: null;
		limit: null;
		offset: null;
		order: null;
		sort: null;
	};
	history: {
		id: null;
	};
	landpads: {
		limit: null;
		offset: null;
	};
	landpad: {
		id: null;
	};
	launches: {
		find: null;
		limit: null;
		offset: null;
		order: null;
		sort: null;
	};
	launchesPast: {
		find: null;
		limit: null;
		offset: null;
		order: null;
		sort: null;
	};
	launchesPastResult: {
		find: null;
		limit: null;
		offset: null;
		order: null;
		sort: null;
	};
	launchesUpcoming: {
		find: null;
		limit: null;
		offset: null;
		order: null;
		sort: null;
	};
	launch: {
		id: null;
	};
	launchLatest: {
		offset: null;
	};
	launchNext: {
		offset: null;
	};
	launchpads: {
		limit: null;
		offset: null;
	};
	launchpad: {
		id: null;
	};
	missions: {
		find: null;
		limit: null;
		offset: null;
	};
	missionsResult: {
		find: null;
		limit: null;
		offset: null;
	};
	mission: {
		id: null;
	};
	payloads: {
		find: null;
		limit: null;
		offset: null;
		order: null;
		sort: null;
	};
	payload: {
		id: null;
	};
	roadster: {};
	rockets: {
		limit: null;
		offset: null;
	};
	rocketsResult: {
		limit: null;
		offset: null;
	};
	rocket: {
		id: null;
	};
	ships: {
		find: null;
		limit: null;
		offset: null;
		order: null;
		sort: null;
	};
	shipsResult: {
		find: null;
		limit: null;
		offset: null;
		order: null;
		sort: null;
	};
	ship: {
		id: null;
	};
	posts: {};
	postComments: {
		postID: null;
	};
	jsp_users: {};
	userPosts: {
		userID: null;
	};
}

const linkBuilder = sourceStep<SourceFields, TargetTypes>();